[SPARK-46503][PS][TESTS] Move `test_default_index` to `pyspark.pandas.tests.indexes.*`
### What changes were proposed in this pull request?
Move `test_default_index` to `pyspark.pandas.tests.indexes.*`

### Why are the changes needed?
test code clean up

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44482 from zhengruifeng/ps_test_idx_default.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
