+++ dev/sparktestsupport/modules.py
@@ -723,7 +723,7 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.plot.test_series_plot_plotly",
         "pyspark.pandas.tests.test_categorical",
         "pyspark.pandas.tests.test_config",
-        "pyspark.pandas.tests.test_default_index",
+        "pyspark.pandas.tests.indexes.test_default",
         "pyspark.pandas.tests.window.test_expanding",
         "pyspark.pandas.tests.window.test_expanding_adv",
         "pyspark.pandas.tests.window.test_expanding_error",
@@ -1049,7 +1049,7 @@ pyspark_pandas_connect_part0 = Module(
         "pyspark.pandas.tests.connect.plot.test_parity_series_plot_plotly",
         "pyspark.pandas.tests.connect.test_parity_categorical",
         "pyspark.pandas.tests.connect.test_parity_config",
-        "pyspark.pandas.tests.connect.test_parity_default_index",
+        "pyspark.pandas.tests.connect.indexes.test_parity_default",
         "pyspark.pandas.tests.connect.test_parity_extension",
         "pyspark.pandas.tests.connect.test_parity_frame_spark",
         "pyspark.pandas.tests.connect.test_parity_generic_functions",
