+++ python/pyspark/pandas/tests/connect/indexes/test_parity_default.py
@@ -16,13 +16,15 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_default_index import DefaultIndexTestsMixin
+from pyspark.pandas.tests.indexes.test_default import DefaultIndexTestsMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
 class DefaultIndexParityTests(
-    DefaultIndexTestsMixin, PandasOnSparkTestUtils, ReusedConnectTestCase
+    DefaultIndexTestsMixin,
+    PandasOnSparkTestUtils,
+    ReusedConnectTestCase,
 ):
     @unittest.skip("Test depends on SparkContext which is not supported from Spark Connect.")
     def test_index_distributed_sequence_cleanup(self):
@@ -30,7 +32,7 @@ class DefaultIndexParityTests(
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_default_index import *  # noqa: F401
+    from pyspark.pandas.tests.connect.indexes.test_parity_default import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
